Fix nasty overhead of pthread_get/setspecific on OS X. Closes #7602.
This patch changes the GC to use special inline variants of
pthread_getspecific and pthread_setspecific, using an Apple-internal
TLS key on OS X. This variable is dedicated to JavaScriptCore's
allocator but otherwise seems unused.

It's currently not enabled by default. When using llvm-gcc or clang, it
will still fall back to using regular pthread_getspecific and
setspecific, which is the current (slow) behavior.

While clang on modern OS X (10.7+) supports __thread variables, this
still has an unfortunate amount of overhead because it requires
a call instruction through a register, like:

    a.out[0x100000eef]:  callq  0x100000f36
    a.out[0x100000ef4]:  leaq   293(%rip), %rdi           ; foo
    a.out[0x100000efb]:  movl   %eax, -20(%rbp)
    a.out[0x100000efe]:  callq  *(%rdi)
    a.out[0x100000f00]:  movl   -20(%rbp), %ecx
    a.out[0x100000f03]:  movl   %ecx, (%rax)

on every access. This is a bit better than the direct pthread calls, but
this is still rather unacceptable for the GC which is perfomance
critical in the threaded RTS. GCC 4.7.x is presumably even worse with
__thread since it does a call through libgcc. On the other hand, it
could use global register variables.

Instead we steal a TLS key from JavaScriptCore, which allows us to use
inline operators that compile down to a single `mov`:

    a.out[0x100000f0c]:  callq  0x100000f38
    a.out[0x100000f11]:  movslq %eax, %rax
    a.out[0x100000f14]:  movq   %rax, %gs:752
    a.out[0x100000f1d]:  movq   %gs:752, %rsi

this is rather underdocumented, but the specifics are in ticket #7602
and should specify everything else necessary.

nofib results are:

TODO

Signed-off-by: Austin Seipp <aseipp@pobox.com>